code review / refactoring, notes
code review / refactoring, notes
